% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oc_browse.R
\name{oc_browse}
\alias{oc_browse}
\title{Browse the Open Context archeological database}
\usage{
oc_browse(type = c("countries", "projects", "descriptions"),
  print_url = FALSE, ...)
}
\arguments{
\item{type}{The kind of to be returned. You can chose either
\code{'countries'} to get a data frame of names of countries that have Open
Context datasets, or \code{'projects'} to get a data frame project names,
or \code{'descriptions'} to get a data frame of data attributes that are
widely used in Open Context data sets.}

\item{print_url}{Whether or not to display a message with the URL of the
query.}

\item{...}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
A data frame with additional class \code{oc_dataframe}.
}
\description{
This function returns a data frame of certain types of top level data from
Open Context. You can get either a data frame of countries for which Open
Context has data, project names that have data on Open Context, or a list of
descriptions (Common Standards) of data attributes that are widely used in
Open Context datasets.
}
\examples{
oc_browse("countries")
oc_browse("projects")
}

