% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oc_get.R
\name{oc_get_countries}
\alias{oc_get_countries}
\title{Retrieve data in Open Context by country}
\usage{
oc_get_countries(country, type = c("projects", "locations", "descriptions"))
}
\arguments{
\item{country}{A character vectory of country names or a data frame
returned from \code{\link{oc_browse}}.}

\item{type}{The type of data to return: \code{"projects"},
\code{"locations"}, \code{"descriptions"}. The default is
\code{"projects"}.}
}
\value{
A data frame with the additional class \code{oc_dataframe}.
}
\description{
Given a character vector of one or more countries, or a data frame of
countries returned from \code{\link{oc_browse}}, this function retrieves data
related to those countries. The function can return projects, locations, or
descriptions.
}
\examples{
oc_get_countries("germany", type= "projects")

library(dplyr)
oc_browse(type = "countries") \%>\%
  filter(label == "Turkey") \%>\%
  oc_get_countries(type = "locations")
}

