% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oc_get.R
\name{oc_get_locations}
\alias{oc_get_locations}
\title{Retrieve the names of projects in a given Country and Location}
\usage{
oc_get_locations(country, locations, type = c("projects", "descriptions"))
}
\arguments{
\item{country}{A country name}

\item{locations}{A character vector of locations in that country}

\item{type}{The type of records to return.}
}
\description{
Retrieve the names of projects in a given Country and Location
}
\examples{
oc_get_locations("Turkey", "Ulucak")
}

