% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oc_get.R
\name{oc_get_projects}
\alias{oc_get_projects}
\title{Retrieve data given an Open Context project name}
\usage{
oc_get_projects(project)
}
\arguments{
\item{project}{A character vector of project names}
}
\value{
A data frame of resources associated with the projects.
}
\description{
Retrieve data given an Open Context project name
}
\examples{
oc_get_projects("Kenan Tepe")
}

