% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opencontext-package.r
\docType{package}
\name{opencontext}
\alias{opencontext}
\alias{opencontext-package}
\title{An API client for the Open Context archaeological database}
\description{
An API client for the Open Context archaeological database
}

